Fix image pull after a failure
When resuming from a failed pull writer.Truncate() was not
seeking to the proper position in the file. This caused writes to
happen after the previously written content, instead of at the start
of the file.

Signed-off-by: Daniel Nephin <dnephin@gmail.com>
(cherry picked from commit 91849080758e2c2dee3d7253f7fd0d3a40f0ec19)
Signed-off-by: Stephen J Day <stephen.day@docker.com>