linux/prox: timeout fifo creation
Under certain conditions in the client, the fifo for a container may not
be created. A timeout has been added to this operation to ensure the
shim can recover when the client fails to open the fifos.

Signed-off-by: Stephen J Day <stephen.day@docker.com>
(cherry picked from commit 9754696ff53dd177a343615e6c6d7dbdc94e21a3)
Signed-off-by: Stephen J Day <stephen.day@docker.com>