Fix label being put on snapshot instead of content
The uncompressed label should be placed on content instead
of snapshots. Currently the uncompressed label is getting
passed into as a label option for apply, which is used to commit
snapshots. The content is already committed, but after apply
the uncompressed diff digest is verified and can be reliably used
to update the content with the label.

Signed-off-by: Derek McGowan <derek@mcgstyle.net>