Edit Containerd license info so GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit removes the LICENSE.docs file and updates the README to
refer readers directly to the CC-BY-4.0 license on the Creative Commons
site. It also updates the copyright reference to "The containerd
Authors" and removes a reference to copyright year in the LICENSE file
(per a recommendation from the CNCF).

This change allows Licensee to successfully identify the license
type of Containerd's codebase as Apache 2.0.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>