Add a passthrough opt for spec configs
This adds a spec config that pipes the spec to binaies located on the
system and in turn, receives a spec back.  It also provides a harness to
create plugins that use existing spec opts.

This sample plugin makes every container greet you with a "hello" when
installed.

```go
package main

import (
	"github.com/containerd/containerd/oci"
	"github.com/containerd/containerd/oci/passthrough"
)

func main() {
	passthrough.Main("greeter", oci.WithProcessArgs("echo", "hello"))
}
```

The default directory for these are `/opt/containerd/bin`.

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>