Implement GPGVault class to hold private keys needed for decryption
Implement a GPGVault class that can hold multiple private keys
of which one or multiple may be needed for decryption of layers.

This class is used when an upper layer, such as Kubernets, instructs
containerd to pull an image from a repository and it is now known
which one(s) of multiple keys are needed to decrypt the image.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>