Add blob mount support to remotes/docker
Client code that uses containerd to manipulate images in a registry are
missing this very important feature (mounting blobs from another
repository). This also opens the door to leverage that when we'll track layers
origin in the local store, to avoid duplicating layers in the registry
(and make Push more efficient).

Signed-off-by: Simon Ferquel <simon.ferquel@docker.com>