Windows:Tests use cmd not powershell
Signed-off-by: John Howard <jhoward@microsoft.com>

powershell doesn't exist in the nanoserver image on RS5 (it did on RS1).
Hence porting to use cmd.exe instead. Note that sleep.exe doesn't exist
on Windows, and isn't a built-in function to cmd.exe, so instead using
a 'ping localhost -n....' hack which simulates a sleep for x seconds.