Rename "content" gRPC plugin to "content-grpc"
The ability to disable plugins introduced in #2105 assumes there is
only one plugin with a given ID (it break's from ranging over registered
plugins after finding an ID match). However, nothing about plugin
registration enforces uniqueness of the ID. Until that is changed,
it is useful to distinguish between the GRPCPlugin and ContentPlugin
which are both named "content".

Signed-off-by: Jared Cordasco <jcordasc@coglib.com>