Unmount rootfs with separate Remove() in bundle
This ensures that a container does not have a mounted rootfs in the
bundle directory before RemoveAll is called.  Having the rootfs removed
first with a Remove ensures that the directory is not mounted and empty
before the bundle directory is removed.

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>
(cherry picked from commit ed308ea1e6f7be0965ddb3afa7df3b86387ba4e9)
Signed-off-by: Wei Fu <fuweid89@gmail.com>