Add user agent header to all requests
Currently the user agent is only being used on the initial
resolve request, then switching to the default user agent.
This ensures the correct user agent is always used. There is
a larger fix in progress which does this is a cleaner way, but
the scope of this change is fixing the user agent issue.

Signed-off-by: Derek McGowan <derek@mcgstyle.net>
(cherry picked from commit bb008728001199665d0a6a484d064372487e7150)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>