Add namespace query parameter for mirrors
Mirrored registries are design to serve content from upstreams.
However, the mirror hostname will usually not match the hostname
of the upstream, requiring the mirror to only use a single
upstream or use its own pattern matching to determine the upstream.
To solve this issue, the client will pass along the namespace which
is being used for the request, allowing mirrors to easily map
to multiple upstreams. This query parameter can safely be ignored
if multiple upstreams are not supported.

Signed-off-by: Derek McGowan <derek@mcgstyle.net>