Release containerd for multiple architectures
Current distributions are not able to keep up with the pace that
projects like containerd are releasing, preventing users using
new features until a new distribution is released, somebody
creates "unofficial" packages or building locally the project.

Containerd provides binaries in a tarball that are easy to use,
but currently there are only amd64 binaries.

This patch adds a script that allows to cross-compile and release
containerd for multiple architectures.

Signed-off-by: Antonio Ojea <antonio.ojea.garcia@gmail.com>