Enable snapshotter to make snapshots without pulling the contents
This commit allows snapshotter to prepare snapshots as well as to prevent the
client from downloading the layer's contents. This hopefully enables to reduce
time for pulling the image. The client passes basic information of the image
layer(ref, digest) to the snapshotter as labels and the snapsohtter can use or
simply ignore it.

This commit also includes changes on ctr command to make using this
functionality easier.

Signed-off-by: Kohei Tokunaga <kouhei.tokunaga.uy@hco.ntt.co.jp>